# be/openapi.yaml
openapi: 3.0.3
info:
  title: cocono-chat backend API
  version: 0.1.0
  description: |
    REST API for cocono-chat (milestones 1–2: accounts + multi-device).

    ## Encoding conventions

    - All binary material (public keys, AES keys, signatures, nonces) is encoded as
      **base64url** (RFC 4648 §5, no padding).
    - Ed25519 public keys are the **raw 32-byte** key (`crypto.subtle.exportKey('raw')`
      in the browser).
    - Signatures are computed over the UTF-8 bytes of the **canonical JSON** of the
      signed object: object keys sorted lexicographically, no whitespace
      (e.g. `canonical({u,p,a,d,t})` → `{"a":...,"d":...,"p":...,"t":...,"u":...}`).
    - Signed payloads (signup, device enrollment) include a client timestamp `t`
      (epoch seconds) that must be within `SIGNED_PAYLOAD_MAX_AGE_SEC` (default 5
      minutes); each signature is additionally de-duplicated in Redis, so captured
      payloads cannot be replayed.

    ## Authentication

    There are no passwords. Login is a challenge-response proof of possession of a
    registered device's Ed25519 private key:

    1. `POST /api/auth/challenge` → one-time nonce `n` (Redis-backed, TTL, single-use).
       A nonce is issued for ANY username/device pair — unknown pairs get a nonce that
       simply never verifies, so the endpoint does not reveal which accounts exist.
    2. Client signs the nonce string (UTF-8 bytes) with the device key.
...
       `Authorization: Bearer <token>`.

...

...

...

...
servers:
...
paths:
...
  /api/share/hit:
    post:
      tags: [accounts]
      operationId: shareHit
      summary: Report that this account opened someone's share link
      description: |
        The "seen" half of share-link attribution (docs/SHARES.md): an account
        that ALREADY existed opened `/?chat=<o>`. Requires a session — the fact
        being stored is a relation between two accounts, and an anonymous
        endpoint here would be a free write-anywhere graph spoofer. The write
        is an upsert on the (owner, viewer) pair, so repeated clicks bump a
        counter instead of adding rows; rate limited per account (`sharehit`).

        Opening your OWN link is not an attribution event: `{recorded:false}`.
        Nothing about any conversation is recorded.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [o]
              properties:
                o:
                  allOf:
                    - $ref: '#/components/schemas/Username'
                  description: The link owner (the account that shared).
            example: { o: alice }
      responses:
        '200': { description: '`{recorded: true|false, o}` — false for your own link.' }
        '400': { description: 'Malformed username.' }
        '401': { description: 'Missing or invalid token.' }
        '404': { description: 'The link owner no longer exists.' }
        '429':
          $ref: '#/components/responses/RateLimited'
  /api/me/share-link:
    get:
      tags: [accounts]
      operationId: myShareLink
      summary: Your own share-link path and what it has produced (counts only)
      description: |
        `{ ul, path, created, clicked }` — the deep link to hand out, how many
        accounts were CREATED from it, and how many accounts have opened it.
        Counts only: the app never learns WHO, that view belongs to the admin
        panel (docs/SHARES.md).
      security:
        - bearerAuth: []
      responses:
        '200': { description: 'Link path + created/clicked counts.' }
        '401': { description: 'Missing or invalid token.' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: HS256 JWT issued by /api/auth/verify.
  responses:
    RateLimited:
      description: 'Too many requests (`error: rate_limited`).'
      headers:
        Retry-After:
          description: Seconds to wait before retrying.
          schema:
            type: integer
            minimum: 1
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Username:
      type: string
      description: 5-64 characters of `[a-zA-Z0-9_-]`; normalised to lowercase at signup, case-insensitively unique.
      pattern: '^[a-zA-Z0-9_-]{4,64}$'
      example: alice
...
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Machine-readable error code (documented per endpoint).
          example: invalid_username
        message:
          type: string
          description: Human-readable description.
          example: Username must be 5+ chars of [a-zA-Z0-9_-]